refactor: move producers and client connector in common module
