--- src/oneid/oneid-common/connector/src/main/java/it/pagopa/oneid/common/DummyClient.java
@@ -1,6 +1,0 @@
-package it.pagopa.oneid.common;
-
-public class DummyClient {
-
-  public static final String CLIENT_NAME = "DummyClient";
-}
