+++ src/oneid/oneid-common/connector/src/main/java/it/pagopa/oneid/common/connector/ClientConnector.java
@@ -1,6 +1,6 @@
-package it.pagopa.oneid.connector;
+package it.pagopa.oneid.common.connector;
 
-import it.pagopa.oneid.common.Client;
+import it.pagopa.oneid.common.model.Client;
 import java.util.ArrayList;
 import java.util.Optional;
