+++ src/oneid/oneid-common/model/src/main/java/it/pagopa/oneid/common/model/Client.java
@@ -1,4 +1,4 @@
-package it.pagopa.oneid.common;
+package it.pagopa.oneid.common.model;
 
 import jakarta.validation.constraints.NotNull;
 import java.util.List;
