+++ src/oneid/oneid-common/model/src/main/java/it/pagopa/oneid/common/model/exception/enums/ErrorCode.java
@@ -1,4 +1,4 @@
-package it.pagopa.oneid.common.exception.enums;
+package it.pagopa.oneid.common.model.exception.enums;
 
 import lombok.Getter;
