+++ src/oneid/oneid-ecs-core/src/main/java/it/pagopa/oneid/common/model/SAMLUtilsConstants.java
@@ -1,4 +1,4 @@
-package it.pagopa.oneid.common;
+package it.pagopa.oneid.common.model;
 
 import org.eclipse.microprofile.config.inject.ConfigProperty;
