+++ src/oneid/oneid-ecs-core/src/main/java/it/pagopa/oneid/exception/IDPSSOEndpointNotFoundException.java
@@ -1,7 +1,7 @@
 package it.pagopa.oneid.exception;
 
-import it.pagopa.oneid.common.exception.OneIdentityException;
-import it.pagopa.oneid.common.exception.enums.ErrorCode;
+import it.pagopa.oneid.common.model.exception.OneIdentityException;
+import it.pagopa.oneid.common.model.exception.enums.ErrorCode;
 
 public class IDPSSOEndpointNotFoundException extends OneIdentityException {
