+++ src/oneid/oneid-ecs-core/src/main/java/it/pagopa/oneid/service/SAMLServiceImpl.java
@@ -5,9 +5,9 @@ import static it.pagopa.oneid.service.utils.SAMLUtilsExtended.buildNameIdPolicy;
 import static it.pagopa.oneid.service.utils.SAMLUtilsExtended.buildRequestedAuthnContext;
 import static it.pagopa.oneid.service.utils.SAMLUtilsExtended.generateSecureRandomId;
 import io.quarkus.logging.Log;
-import it.pagopa.oneid.common.SAMLUtilsConstants;
-import it.pagopa.oneid.common.exception.OneIdentityException;
-import it.pagopa.oneid.common.exception.SAMLUtilsException;
+import it.pagopa.oneid.common.model.SAMLUtilsConstants;
+import it.pagopa.oneid.common.model.exception.OneIdentityException;
+import it.pagopa.oneid.common.model.exception.SAMLUtilsException;
 import it.pagopa.oneid.exception.GenericAuthnRequestCreationException;
 import it.pagopa.oneid.exception.IDPSSOEndpointNotFoundException;
 import it.pagopa.oneid.exception.SAMLResponseStatusException;
